change size_t to int

+++ 9-binary_tree_height.c
@@ -7,7 +7,7 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	size_t height = 0;
+	int height = 0;
 
 	if (tree == NULL)
 	{
